Filled in all of the network processing fields and functions for area detectors.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@488 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_image.h
@@ -58,7 +58,7 @@ typedef struct {
 	long framesize[2];
 	long image_format;
 	long pixel_order;
-	long bytes_per_pixel;
+	double bytes_per_pixel;
 
 	size_t header_length;
 	void *header_data;
